# [Internal] Binary Encoding: Add performance tests.